requirements.txt
scrapy==2.11.2
scrapy-playwright==0.0.40
playwright==1.44.0
tldextract==5.1.2
requests==2.32.3
tenacity==8.3.0
pyyaml==6.0.1
python-dotenv==1.0.1
rich==13.7.1
lxml==5.2.2
cssselect==1.2.0
